issue 37 update added bottom nav bar to summary/crossblock/wishlist fragments updated ui in data summary

+++ app/src/main/java/org/phenoapps/intercross/adapters/SummaryAdapter.kt
@@ -60,7 +60,7 @@ class SummaryAdapter:
 
                 labelText = data.label
 
-                countValue = data.value
+                countValue = data.value.toInt()
 
             }
         }
